fix(nestjs-bus): rabbitmq module initialization & implement basic tests for rabbitmq module
affects: @atlantis-lab/nestjs-bus

ISSUES CLOSED: #190
